Add drain command
When we need to reassign replicas from a broker to other brokers,
say, when the broker is dead, we used to use `reassign` command,
which reassign all replicas randomly, whereas replicas on healthy
brokers might be moved around. This is unefficient.

In this PR, I bring into a new command `drain`, which essentially
empty a specific broker and assign its replicas to a list of
healthy ones. Replicas on healthy brokers will not be moved around.

Usage:
drain <broker id> [topic] [--brokers <ids>]

Tests are also be brought into this repo, start from covering the
new command.