Obtain the sequence of restarting kafkat brokers with minimum cost
Goal:
This PR includes the commands and algorithms that help calculating a sequence
for restarting the kafkat brokers. The algorithm tries to minimize the restart
cost for the sequence using a greedy algorithm.

Defination of Cost:
The cost of each leader replica is defined as:
cost_of_leader_restart($broker, $partition) = replication_factor($partition) -
number_of_replica_restarted($partition)

And the cost of each broker is defined as:
cost_of_restart($broker) =  sum of cost_of_leader_restart($broker, $partition)
for all leaders on $brokers

Commands:
cluster-restart help                Print Help and exit
cluster-restart reset               Clean up the restart state
cluster-restart start               Initialize the cluster-restart session for the cluster
cluster-restart next                Calculate the next broker to restart based on the current state
cluster-restart good <broker>       Mark this broker as successfuly restarted
cluster-restart log                 Print the state of the brokers
cluster-restart restore <file>      Start a new session and restore the state defined in that file

Test:
We implemented spec tests for each of the class in this script.
In addition, we also ran this script against a staging zookeeper, and it worked correctly.