Add rubocop style check, clean up, file renames
Based the rubocop configuration off of the shopify ruby styleguide which
is a pretty nice starting point and added some customizations that fit
well with the existing project.  Significant work was done to bring the
code base into compliance with the style guide.  Standardized file
naming with snake case instead a mix of snake and dashes.  In addition,
factory_girl was replaced by factory_bot and the factories updated.