Final style cleanup
Fixes a handful of style issues that the editor integration did not
catch.  Also rather than ignore the catchall rescue in the config
handler, opt to remove that and actually expose any error that may be
present.  We can make a determination on how to handle individual
exceptions. Add rubocop to the travis check and update ruby versions